+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `Innmind\AMQP\Factory::make()` timeout argument is now expressed via `Innmind\TimeContinuum\Period`
 - `Innmind\AMQP\Model\Basic\Message` expiration is now expressed via `Innmind\TimeContinuum\Period`
 - `Innmind\AMQP\Failure` is now an exception that wraps each possible failure object
+- `Innmind\AMQP\Client::run()` now returns an `Innmind\Immutable\Attempt`
 
 ### Fixed